# Daily Perf Improver: VSIDS Heap and Activity Management Optimizations (Round 2)